Move shutdown logic to `main.rs`
