+++ bin/inx-chronicle/src/main.rs
@@ -11,7 +11,6 @@ mod config;
 mod launcher;
 #[cfg(feature = "metrics")]
 mod metrics;
-mod shutdown;
 #[cfg(all(feature = "stardust", feature = "inx"))]
 mod stardust_inx;
 
@@ -40,9 +39,35 @@ async fn startup(scope: &mut RuntimeScope) -> Result<(), Box<dyn Error + Send + 
     let launcher_addr = scope.spawn_actor_unsupervised(Launcher).await;
 
     spawn_task("shutdown listener", async move {
-        shutdown::shutdown_signal_listener().await;
+        shutdown_signal_listener().await;
         launcher_addr.abort().await;
     });
 
     Ok(())
+}
+
+async fn shutdown_signal_listener() {
+    #[cfg(unix)]
+    {
+        use futures::future;
+        use tokio::signal::unix::{signal, Signal, SignalKind};
+
+        // Panic: none of the possible error conditions should happen.
+        let mut signals = vec![SignalKind::interrupt(), SignalKind::terminate()]
+            .iter()
+            .map(|kind| signal(*kind).unwrap())
+            .collect::<Vec<Signal>>();
+        let signal_futs = signals.iter_mut().map(|signal| Box::pin(signal.recv()));
+        let (signal_event, _, _) = future::select_all(signal_futs).await;
+
+        if signal_event.is_none() {
+            panic!("Shutdown signal stream failed, channel may have closed.");
+        }
+    }
+    #[cfg(not(unix))]
+    {
+        if let Err(e) = tokio::signal::ctrl_c().await {
+            panic!("Failed to intercept CTRL-C: {:?}.", e);
+        }
+    }
 }
